+++ src/app/d2-vendors/vendors.tsx
@@ -5,7 +5,9 @@ import {
   DestinyVendorComponent,
   DestinyVendorSaleItemComponent,
   DestinyVendorsResponse,
-  DestinyVendorGroup
+  DestinyVendorGroup,
+  DestinyVendorDefinition,
+  BungieMembershipType
   } from 'bungie-api-ts/destiny2';
 import * as React from 'react';
 import { DestinyAccount } from '../accounts/destiny-account.service';
@@ -21,6 +23,8 @@ import { DestinyTrackerServiceType } from '../item-review/destiny-tracker.servic
 import { fetchRatingsForVendors } from './vendor-ratings';
 import { Subscription } from 'rxjs/Subscription';
 import { D2StoreServiceType, D2Store } from '../inventory/store-types';
+import { VendorItem } from './vendor-item';
+import { D2ReviewDataCache } from '../destinyTrackerApi/d2-reviewDataCache';
 
 interface Props {
   $scope: IScope;
@@ -99,6 +103,7 @@ export default class Vendors extends React.Component<Props, State> {
 
   render() {
     const { defs, vendorsResponse, trackerService, ownedItemHashes } = this.state;
+    const { account } = this.props;
 
     if (!vendorsResponse || !defs) {
       // TODO: loading component!
@@ -115,6 +120,7 @@ export default class Vendors extends React.Component<Props, State> {
             vendorsResponse={vendorsResponse}
             trackerService={trackerService}
             ownedItemHashes={ownedItemHashes}
+            account={account}
           />
         )}
 
@@ -128,13 +134,15 @@ function VendorGroup({
   group,
   vendorsResponse,
   trackerService,
-  ownedItemHashes
+  ownedItemHashes,
+  account
 }: {
   defs: D2ManifestDefinitions;
   group: DestinyVendorGroup;
   vendorsResponse: DestinyVendorsResponse;
   trackerService?: DestinyTrackerServiceType;
   ownedItemHashes?: Set<number>;
+  account: DestinyAccount;
 }) {
   const groupDef = defs.VendorGroup.get(group.vendorGroupHash);
   return (
@@ -143,6 +151,7 @@ function VendorGroup({
       {group.vendorHashes.map((h) => vendorsResponse.vendors.data[h]).map((vendor) =>
         <Vendor
           key={vendor.vendorHash}
+          account={account}
           defs={defs}
           vendor={vendor}
           itemComponents={vendorsResponse.itemComponents[vendor.vendorHash]}
@@ -163,7 +172,8 @@ function Vendor({
   sales,
   trackerService,
   ownedItemHashes,
-  currencyLookups
+  currencyLookups,
+  account
 }: {
   defs: D2ManifestDefinitions;
   vendor: DestinyVendorComponent;
@@ -176,6 +186,7 @@ function Vendor({
   currencyLookups: {
     [itemHash: number]: number;
   };
+  account: DestinyAccount;
 }) {
   const vendorDef = defs.Vendor.get(vendor.vendorHash);
   if (!vendorDef) {
@@ -202,12 +213,45 @@ function Vendor({
       <VendorItems
         defs={defs}
         vendorDef={vendorDef}
-        sales={sales}
-        itemComponents={itemComponents}
+        vendorItems={getVendorItems(account, defs, vendorDef, trackerService, itemComponents, sales)}
         trackerService={trackerService}
         ownedItemHashes={ownedItemHashes}
         currencyLookups={currencyLookups}
       />
     </div>
   );
+}
+
+export function getVendorItems(
+  account: DestinyAccount,
+  defs: D2ManifestDefinitions,
+  vendorDef: DestinyVendorDefinition,
+  trackerService?: DestinyTrackerServiceType,
+  itemComponents?: DestinyItemComponentSetOfint32,
+  sales?: {
+    [key: string]: DestinyVendorSaleItemComponent;
+  }
+) {
+  const reviewCache: D2ReviewDataCache | undefined = trackerService ? trackerService.getD2ReviewDataCache() : undefined;
+
+  if (sales && itemComponents) {
+    const components = Object.values(sales);
+    return components.map((component) => new VendorItem(
+      defs,
+      vendorDef,
+      vendorDef.itemList[component.vendorItemIndex],
+      reviewCache,
+      component,
+      itemComponents
+    ));
+  } else if (vendorDef.returnWithVendorRequest) {
+    // If the sales should come from the server, don't show anything until we have them
+    return [];
+  } else {
+    return vendorDef.itemList.filter((i) =>
+      !i.exclusivity ||
+      i.exclusivity === BungieMembershipType.All ||
+      i.exclusivity === account.platformType
+    ).map((i) => new VendorItem(defs, vendorDef, i, reviewCache));
+  }
 }